pricing/requirements.txt
crewai
langchain_openai
python-decouple
textwrap3
duckduckgo-search
langchain-community
tavily-python
unstructured
